+++ exts/sl.sensor.camera/docs/CHANGELOG.md
@@ -2,6 +2,9 @@
 
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
+## [4.2.1]
+- Fix streaming of 4mm camera models (ZED X and ZED XM).
+
 ## [4.2.0]
 - Add the option to change streaming bitrate and chunksize when streaming over network.
 - Add the option to stream over both IPC and network simultaneously.